Move components closer
