+++ package.json
@@ -14,8 +14,8 @@
   },
   "devDependencies": {
     "@playwright/test": "^1.25.0",
-    "@sveltejs/adapter-cloudflare": "1.0.0-next.31",
-    "@sveltejs/kit": "1.0.0-next.411",
+    "@sveltejs/adapter-cloudflare": "^1.0.0-next.31",
+    "@sveltejs/kit": "^1.0.0-next.415",
     "@tailwindcss/forms": "^0.5.2",
     "@tailwindcss/line-clamp": "^0.4.0",
     "@tailwindcss/typography": "^0.5.4",
