+++ src/routes/ChocoSection.svelte
